+++ VirtualH89/Src/H47Drive.cpp
@@ -72,20 +72,20 @@ H47Drive::step(bool direction)
 }
 
 BYTE
-H47Drive::readData(unsigned int pos)
+H47Drive::readData(unsigned long pos)
 {
 
     return 0;
 }
 
 void
-H47Drive::writeData(unsigned int pos, BYTE data)
+H47Drive::writeData(unsigned long pos, BYTE data)
 {
 
 }
 
 BYTE
-H47Drive::readSectorData(BYTE sector, unsigned int pos)
+H47Drive::readSectorData(BYTE sector, unsigned long pos)
 {
 
     return 0;
